--- src/wai_data_tools/scripts/__init__.py
@@ -1,1 +1,0 @@
-"""Init module."""
